+++ src/modules.d.ts
@@ -1,3 +1,0 @@
-declare module '#XMLHttpRequest.cjs' {
-  export const XMLHttpRequest: typeof globalThis.XMLHttpRequest;
-}