+++ test/integration/node-es6/index.js
@@ -1,5 +1,10 @@
 import { createSQLiteThread, createHttpBackend } from 'sqlite-wasm-http';
+import sqlite from 'sqlite-wasm-http/sqlite3.js';
+
 import './setup.js';
+
+if (typeof sqlite !== 'function')
+  throw new Error('Importing sqlite3 subpath export failed');
 
 (async function main() {
   const httpBackend = createHttpBackend({
@@ -32,7 +37,7 @@ import './setup.js';
   });
 
   if (rows.length < 6)
-    throw new Error(('Could not retrieve all rows'));
+    throw new Error('Could not retrieve all rows');
 
   await db('close', {});
   db.close();