+++ test/integration/node-ts-esm/modules.d.ts
@@ -1,0 +1,1 @@
+declare module '@mmomtchev/web-worker';